requirements.txt
elevenlabs==1.51.0
fal-client==0.5.9
flask==3.0.2
flask-socketio==5.3.6
flask-cors
python-dotenv==1.0.0
websockets==12.0
python-socketio
eventlet

